Cleaned up imports and logger usage
